requirements.txt
python-Levenshtein>=0.21.1
rapidfuzz>=3.6.1
pdfplumber>=0.10.3
tabulate>=0.9.0 